requirements.txt
functions-framework==3.*
requests==2.26.0
pandas==1.4.0
google-cloud-storage==1.42.3